refactor: move board.service from separate folder
