--- src/resources/boards/board.service/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./board.service')
